farm/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiogram-media-group==0.5.1
aiohttp==3.9.0
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.6.0
APScheduler==3.10.4
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.0
Deprecated==1.2.14
firebase-admin==6.3.0
frozenlist==1.3.3
gcloud==0.18.3
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
jwcrypto==1.5.1
magic-filter==1.0.12
msgpack==1.0.7
multidict==6.0.4
numpy==1.26.3
oauth2client==4.1.3
openai==0.27.4
pandas==2.2.0
ply==3.11
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.4.1
pydantic_core==2.10.1
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-dateutil==2.8.2
python-jwt==4.1.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.3.23
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
smop==0.41
tiktoken==0.3.3
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
wrapt==1.16.0
yarl==1.8.2
YooMoney==0.1.0
sulguk[cli]
